pyproject.toml
[project]
name = "scpi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "psutil>=7.2.2",
    "pyserial>=3.5",
    "pyvisa>=1.16.2",
    "pyvisa-py>=0.8.1",
    "zeroconf>=0.148.0",
]
